fix: 移除platform和bussiness逻辑代码
* fix issue #720

* 移除platform以及bussiness的相关代码逻辑
