+++ service/main_test.go
@@ -48,7 +48,6 @@ import (
 	"github.com/polarismesh/polaris/namespace"
 	"github.com/polarismesh/polaris/plugin"
 	_ "github.com/polarismesh/polaris/plugin/auth/defaultauth"
-	_ "github.com/polarismesh/polaris/plugin/auth/platform"
 	_ "github.com/polarismesh/polaris/plugin/cmdb/memory"
 	_ "github.com/polarismesh/polaris/plugin/discoverevent/local"
 	_ "github.com/polarismesh/polaris/plugin/discoverstat/discoverlocal"
@@ -78,7 +77,6 @@ const (
 	tblRateLimitRevision      = "ratelimit_revision"
 	tblCircuitBreaker         = "circuitbreaker_rule"
 	tblCircuitBreakerRelation = "circuitbreaker_rule_relation"
-	tblPlatform               = "platform"
 	tblNameL5                 = "l5"
 	tblNameRoutingV2          = "routing_config_v2"
 	tblClient                 = "client"
